+++ htdocs/api/index.php
@@ -33,6 +33,10 @@ use Slim\Routing\RouteCollectorProxy;
 use Twig\Environment;
 use Twig\Loader\FilesystemLoader;
 
+const STAT_TYPE_ALL             = 'all';
+const STAT_TYPE_CUSTOMER_FACING = 'customerFacing';
+const STAT_TYPE_PROACTIVE       = 'proactive';
+const statsTypes                = [STAT_TYPE_ALL, STAT_TYPE_CUSTOMER_FACING, STAT_TYPE_PROACTIVE];
 require_once __DIR__ . '/../config.inc.php';
 global $cfg;
 require_once($cfg["path_dbe"] . "/DBEQuotation.inc.php");
@@ -106,12 +110,25 @@ $app->group(
                         return $response->withStatus(400);
                     }
                 }
-                $params      = [
+                $params    = [
                     ["type" => "i", "value" => $args['customerId']],
                     ["type" => "s", "value" => $startDate->format(DATE_MYSQL_DATE)],
                     ["type" => "s", "value" => $endDate->format(DATE_MYSQL_DATE)],
                 ];
-                $isBreakDown = isset($queryParams['breakDown']);
+                $typeWhere = '';
+                $typeParam = [];
+                if (isset($queryParams['type'])) {
+                    $type = $queryParams['type'];
+                    if (!in_array($type, statsTypes)) {
+                        $typesString = implode(',', statsTypes);
+                        $response->getBody()->write(
+                            json_encode(["error" => "The type must be one of {$typesString}"])
+                        );
+                        return $response->withStatus(400);
+                    }
+                    $typeWhere   = " and pro_hide_from_customer_flag = ?";
+                    $typeParam[] = ["type" => "s", "value" => $type == STAT_TYPE_CUSTOMER_FACING ? 'N' : 'Y'];
+                }
                 $query       = "select SUM(1) AS raised,
     SUM(pro_status IN ('F' , 'C')) AS `fixed`,
     AVG(if(pro_status IN ('F', 'C'),problem.`pro_responded_hours`,null)) AS responseTime,
@@ -195,7 +212,9 @@ WHERE
        
         problem.pro_custno = ?
   and initial.caa_date between ? and ?
+  {$typeWhere}
         AND pro_priority <= 4";
+                $isBreakDown = isset($queryParams['breakDown']);
                 if ($isBreakDown) {
 
                     $params = [
@@ -259,7 +278,9 @@ FROM
     customer.`slaFixHoursP4` AS fixSLA
   FROM
     customer
-  WHERE customer.`cus_custno` = ?) priorities
+  WHERE customer.`cus_custno` = ?
+      
+      ) priorities
   LEFT JOIN
     (SELECT
       pro_priority AS priority,
@@ -429,9 +450,11 @@ FROM
       AND initial.caa_date BETWEEN ?
       AND ?
       AND pro_priority <= 4
+      {$typeWhere}
     GROUP BY pro_priority) stats
     ON stats.priority = priorities.priority";
                 }
+                $params = array_merge($params, $typeParam);
                 /** @var $db dbSweetcode */ global $db;
                 $statement = $db->preparedQuery($query, $params);
                 if ($isBreakDown) {
@@ -819,17 +842,31 @@ ORDER BY COUNT DESC";
                         return $response->withStatus(400);
                     }
                 }
-                $params      = [
+                $params    = [
                     ["type" => "s", "value" => $startDate->format(DATE_MYSQL_DATE)],
                     ["type" => "s", "value" => $endDate->format(DATE_MYSQL_DATE)],
                 ];
+                $typeWhere = '';
+                $typeParam = [];
+                if (isset($queryParams['type'])) {
+                    $type = $queryParams['type'];
+                    if (!in_array($type, statsTypes)) {
+                        $typesString = implode(',', statsTypes);
+                        $response->getBody()->write(
+                            json_encode(["error" => "The type must be one of {$typesString}"])
+                        );
+                        return $response->withStatus(400);
+                    }
+                    $typeWhere   = " and pro_hide_from_customer_flag = ?";
+                    $typeParam[] = ["type" => "s", "value" => $type == STAT_TYPE_CUSTOMER_FACING ? 'N' : 'Y'];
+                }
                 $isBreakDown = isset($queryParams['breakDown']);
-                $query       = 'SELECT
+                $query       = "SELECT
     SUM(1) AS raised,
-    SUM(pro_status IN("F","C")) AS `fixed`,
-  AVG(if(problem.pro_priority = 1 and pro_status IN ("F","C") ,problem.`pro_responded_hours`, null)) AS responseTime,
+    SUM(pro_status IN(\"F\",\"C\")) AS `fixed`,
+  AVG(if(problem.pro_priority = 1 and pro_status IN (\"F\",\"C\") ,problem.`pro_responded_hours`, null)) AS responseTime,
   AVG(
-   IF(pro_status IN ("F","C"),   
+   IF(pro_status IN (\"F\",\"C\"),   
    problem.`pro_responded_hours` < 
     CASE
       problem.`pro_priority`
@@ -845,11 +882,11 @@ ORDER BY COUNT DESC";
     END,
     NULL) 
   ) AS slaMet,
-    AVG(IF(pro_status IN ("F","C"), openHours < 8, NULL)) AS closedWithin8Hours,
-    AVG(IF(pro_status ="C" and pro_hide_from_customer_flag <> "Y",problem.`pro_reopened_date` IS NOT NULL, NULL)) AS reopened,
-    AVG(IF(pro_status IN ("F","C"), problem.`pro_chargeable_activity_duration_hours`,NULL)) AS avgChargeableTime,
-    AVG(IF(pro_status IN ("F","C"), problem.pro_working_hours,NULL)) AS avgTimeAwaitingCNC,
-    AVG(IF(pro_status IN ("F","C"), openHours,NULL)) AS avgTimeFromRaiseToFixHours
+    AVG(IF(pro_status IN (\"F\",\"C\"), openHours < 8, NULL)) AS closedWithin8Hours,
+    AVG(IF(pro_status =\"C\" and pro_hide_from_customer_flag <> \"Y\",problem.`pro_reopened_date` IS NOT NULL, NULL)) AS reopened,
+    AVG(IF(pro_status IN (\"F\",\"C\"), problem.`pro_chargeable_activity_duration_hours`,NULL)) AS avgChargeableTime,
+    AVG(IF(pro_status IN (\"F\",\"C\"), problem.pro_working_hours,NULL)) AS avgTimeAwaitingCNC,
+    AVG(IF(pro_status IN (\"F\",\"C\"), openHours,NULL)) AS avgTimeFromRaiseToFixHours
 FROM
   problem
   LEFT JOIN callactivity initial
@@ -858,9 +895,10 @@ FROM
   JOIN customer
     ON problem.`pro_custno` = customer.`cus_custno`
 WHERE  caa_date between ? and ?
-  AND pro_priority < 4';
+ {$typeWhere}
+  AND pro_priority < 4";
                 if ($isBreakDown) {
-                    $query = 'SELECT
+                    $query = "SELECT
   p.priority, data.avgChargeableTime, raised, fixed, responseTime, slaMet, closedWithin8Hours, reopened, avgChargeableTime, avgTimeAwaitingCNC, avgTimeFromRaiseToFixHours
 FROM
   (SELECT
@@ -881,18 +919,18 @@ FROM
     (SELECT
       pro_priority AS priority,
       SUM(1) AS raised,
-      SUM(pro_status IN ("F", "C")) AS `fixed`,
+      SUM(pro_status IN (\"F\", \"C\")) AS `fixed`,
       AVG(
         IF(
           problem.pro_priority = 1
-          AND pro_status IN ("F", "C"),
+          AND pro_status IN (\"F\", \"C\"),
           problem.`pro_responded_hours`,
           NULL
         )
       ) AS responseTime,
       AVG(
         IF(
-          pro_status IN ("F", "C"),
+          pro_status IN (\"F\", \"C\"),
           problem.`pro_responded_hours` <
           CASE
             problem.`pro_priority`
@@ -911,36 +949,36 @@ FROM
       ) AS slaMet,
       AVG(
         IF(
-          pro_status IN ("F", "C"),
+          pro_status IN (\"F\", \"C\"),
           openHours < 8,
           NULL
         )
       ) AS closedWithin8Hours,
       AVG(
         IF(
-          pro_status = "C"
-          AND pro_hide_from_customer_flag <> "Y",
+          pro_status = \"C\"
+          AND pro_hide_from_customer_flag <> \"Y\",
           problem.`pro_reopened_date` IS NOT NULL,
           NULL
         )
       ) AS reopened,
       AVG(
         IF(
-          pro_status IN ("F", "C"),
+          pro_status IN (\"F\", \"C\"),
           problem.`pro_chargeable_activity_duration_hours`,
           NULL
         )
       ) AS avgChargeableTime,
       AVG(
         IF(
-          pro_status IN ("F", "C"),
+          pro_status IN (\"F\", \"C\"),
           problem.pro_working_hours,
           NULL
         )
       ) AS avgTimeAwaitingCNC,
       AVG(
         IF(
-          pro_status IN ("F", "C"),
+          pro_status IN (\"F\", \"C\"),
           openHours,
           NULL
         )
@@ -955,11 +993,14 @@ FROM
     WHERE caa_date BETWEEN ?
       AND ?
       AND pro_priority < 5
+      {$typeWhere}
     GROUP BY pro_priority
     ORDER BY pro_priority) data
-    ON data.priority = p.priority';
+    ON data.priority = p.priority";
                 }
                 try {
+
+                    $params = array_merge($params, $typeParam);
                     $statement = $db->preparedQuery($query, $params);
                     if (!$isBreakDown) {
                         $data = $statement->fetch_assoc();